requirements.txt
# Core framework dependencies
PyYAML==6.0.1
pydantic==2.5.0
click==8.1.7
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0

# Mobile automation drivers
uiautomator2==2.16.25
Appium-Python-Client==3.1.0
selenium==4.16.0

# Database dependencies
pymongo==4.6.0
motor==3.3.2
redis==5.0.1

# Testing framework
pytest==7.4.3
pytest-html==4.1.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Reporting
allure-pytest==2.13.2
Jinja2==3.1.2

# Image and video processing
Pillow==10.1.0
opencv-python==4.8.1.78
imageio==2.33.0

# Machine learning (for healing intelligence)
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4

# Utilities
colorama==0.4.6
tqdm==4.66.1
psutil==5.9.6
schedule==1.2.0

# Logging
colorlog==6.8.0

# Configuration management
configparser==6.0.0
jsonschema==4.20.0

# Async support
aiofiles==23.2.1
asyncio==3.4.3

# Development dependencies
pytest-django==4.7.0
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
